% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_rec.R
\name{area_rec}
\alias{area_rec}
\title{Calculate area of a rectangle}
\usage{
area_rec(side1, side2)
}
\arguments{
\item{side1}{A side of a rectangle}

\item{side2}{The other side of the rectangle}
}
\value{
The area of a rectangle
}
