% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exams_check.R
\name{check_answers}
\alias{check_answers}
\title{Checks whether answers make sense}
\usage{
check_answers(answers_in)
}
\arguments{
\item{answers_in}{vector of single choice answers}
}
\value{
logical (TRUE/FALSE)
}
\description{
Checks whether answers make sense
}
\examples{
flag <- check_answers(1:5)
flag
}
