% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{data_dir_get}
\alias{data_dir_get}
\title{Get path of data file}
\usage{
data_dir_get(f_in)
}
\arguments{
\item{f_in}{}
}
\value{
a path
}
\description{
Get path of data file
}
\examples{
data_dir_get('B3-stocks.rds')
}
