% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\name{density2}
\alias{density2}
\title{Density function}
\arguments{
\item{df}{A dataframe}

\item{var1}{A character string of the column name}
}
\value{
Density Plot of input column
}
\description{
Takes a data.frame and column of that data.frame to build a density plot using ggplot.
}
