% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{get_football_games}
\alias{get_football_games}
\title{returns data for footballgames}
\usage{
get_football_games()
}
\value{
dataframe
}
\description{
returns data for footballgames
}
\examples{
get_football_games()
}
