% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{get_SP500_yearly_data}
\alias{get_SP500_yearly_data}
\title{returns data for SP500 (yearly)}
\usage{
get_SP500_yearly_data()
}
\value{
dataframe
}
\description{
returns data for SP500 (yearly)
}
\examples{
get_SP500_yearly_data()
}
