% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{get_SP500_yearly_file}
\alias{get_SP500_yearly_file}
\title{Returns path of file for sp500 yearly}
\usage{
get_SP500_yearly_file()
}
\value{
a path
}
\description{
Returns path of file for sp500 yearly
}
\examples{
get_SP500_yearly_file()
}
