% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_stock_br_data}
\alias{get_stock_br_data}
\title{Get reliable data for yf-BR}
\usage{
get_stock_br_data()
}
\value{
a dataframe
}
\description{
Get reliable data for yf-BR
}
\examples{
get_stock_br_data()
}
