% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_import.R
\name{get_stock_br_file}
\alias{get_stock_br_file}
\title{returns data for stocks br}
\usage{
get_stock_br_file()
}
\value{
a dataframe
}
\description{
returns data for stocks br
}
\examples{
get_stock_br_file()
}
