% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{make_random_answers}
\alias{make_random_answers}
\title{Builds random solutions}
\usage{
make_random_answers(solution, candidates = NA, is_cash = FALSE)
}
\arguments{
\item{solution}{the solution}

\item{candidates}{all canditates for solution}

\item{is_cash}{if solution is cash}
}
\value{
a vector
}
\description{
Builds random solutions
}
\examples{
make_random_answers(199.12, is_cash=TRUE)
}
