% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapIt.R
\name{mapIt}
\alias{mapIt}
\title{Plots a map}
\arguments{
\item{df}{A data.frame}

\item{plotLat}{A character string name of the latitude column from the data.frame}

\item{plotLong}{A character string name of the longitude column from the data.frame}

\item{col}{A character string name of a column for a color ( aes(color=?) ), default to NULL}

\item{latRange}{A tuple vector with the min then max latitudes desired, default  c(-90,90)}

\item{longRange}{A tuple vector with the min then max longitudes desired, default c(-180,180)}
}
\value{
Map plot
}
