% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_exam.R
\name{my_exams2moodle}
\alias{my_exams2moodle}
\title{Produces moodle file from exercises}
\usage{
my_exams2moodle(exercise_files, n_ver, class_name, exerc_name, my_deadline)
}
\arguments{
\item{exercise_files}{path of exercise files}

\item{n_ver}{number of versions}

\item{class_name}{name of class}

\item{exerc_name}{name of exercise}

\item{my_deadline}{deadline time}

\item{stitle}{title of exercise}

\item{n_q}{number of questions}
}
\value{
nothing
}
\description{
Produces moodle file from exercises
}
