% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{print_pretty_df}
\alias{print_pretty_df}
\title{prints a pretty dataframe}
\usage{
print_pretty_df(df)
}
\arguments{
\item{df}{a dataframe}
}
\value{
a pretty html table
}
\description{
prints a pretty dataframe
}
\examples{
print_pretty_df(dplyr::tibble(x = 1, y = 2))
}
