% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_cash}
\alias{sample_cash}
\title{Sample a cash value}
\usage{
sample_cash(min_value = 0, max_value = 100, n_bins = 25)
}
\arguments{
\item{min_value}{min value}

\item{max_value}{max value}

\item{n_bins}{number of bins to use}
}
\value{
a value
}
\description{
Sample a cash value
}
\examples{
sample_cash()
}
