% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_yield_rate}
\alias{sample_yield_rate}
\title{Samples a yield rate}
\usage{
sample_yield_rate(min_value = 0.025, max_value = 0.25, n_bins = 50)
}
\arguments{
\item{min_value}{min value}

\item{max_value}{max value}

\item{n_bins}{number of bins to use}
}
\value{
a yield rate
}
\description{
Samples a yield rate
}
\examples{
sample_yield_rate()
}
