% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top_n.R
\name{top_n}
\alias{top_n}
\title{What is the top n things}
\arguments{
\item{df}{A dataframe}

\item{groupings}{A character string of column to group}

\item{n}{A number to clip count}

\item{count}{A TRUE/FALSE, includes the count column of output data.frame, excludes count column, respectively}
}
\description{
Takes a data.frame and produces the top n of a grouping.
}
